move signal type to own file
