+++ SoundBoard.py
@@ -3,7 +3,7 @@ import threading
 from SoundDict import sounds
 from utils.KeyboardListener import listen_for_key
 from utils.GPIOListener import start_gpio_listener
-from utils.Signals import SignalType
+from utils.SignalType import SignalType
 from utils.GPIOSignalPrinter import process_signals
