+++ utils/Signals.py
@@ -1,20 +1,4 @@
-from enum import Enum, auto
-
-
-class SignalType(Enum):
-    SOUND1 = auto()
-    SOUND2 = auto()
-    SOUND3 = auto()
-    SOUND4 = auto()
-    SOUND5 = auto()
-    SOUND6 = auto()
-    SOUND7 = auto()
-    SOUND8 = auto()
-    SOUND9 = auto()
-
-    REC_BTN = auto()
-    QUIT = auto()
-    STOP_SOUND = auto()
+from SignalType import SignalType
 
 
 class Signals:
